+++ src/main/java/api/m2/file/enums/InvitationStatus.java
@@ -1,0 +1,7 @@
+package api.m2.file.enums;
+
+public enum InvitationStatus {
+    PENDING,
+    ACCEPTED,
+    REJECTED
+}